requirements.txt
flask
opencv-python-headless
torch
torchvision
pillow
numpy